@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 use Psr\Http\Message\MessageInterface;
@@ -12,10 +14,8 @@ final class Message
      * Returns the string representation of an HTTP message.
      *
      * @param MessageInterface $message Message to convert to a string.
-     *
-     * @return string
      */
-    public static function toString(MessageInterface $message)
+    public static function toString(MessageInterface $message): string
     {
         if ($message instanceof RequestInterface) {
             $msg = trim($message->getMethod() . ' '
@@ -52,10 +52,8 @@ final class Message
      *
      * @param MessageInterface $message    The message to get the body summary
      * @param int              $truncateAt The maximum allowed size of the summary
-     *
-     * @return string|null
      */
-    public static function bodySummary(MessageInterface $message, $truncateAt = 120)
+    public static function bodySummary(MessageInterface $message, int $truncateAt = 120): ?string
     {
         $body = $message->getBody();
 
@@ -95,7 +93,7 @@ final class Message
      *
      * @throws \RuntimeException
      */
-    public static function rewindBody(MessageInterface $message)
+    public static function rewindBody(MessageInterface $message): void
     {
         $body = $message->getBody();
 
@@ -112,10 +110,8 @@ final class Message
      * array values, and a "body" key containing the body of the message.
      *
      * @param string $message HTTP request or response to parse.
-     *
-     * @return array
      */
-    public static function parseMessage($message)
+    public static function parseMessage(string $message): array
     {
         if (!$message) {
             throw new \InvalidArgumentException('Invalid message');
@@ -129,7 +125,7 @@ final class Message
             throw new \InvalidArgumentException('Invalid message: Missing header delimiter');
         }
 
-        list($rawHeaders, $body) = $messageParts;
+        [$rawHeaders, $body] = $messageParts;
         $rawHeaders .= "\r\n"; // Put back the delimiter we split previously
         $headerParts = preg_split("/\r?\n/", $rawHeaders, 2);
 
@@ -137,7 +133,7 @@ final class Message
             throw new \InvalidArgumentException('Invalid message: Missing status line');
         }
 
-        list($startLine, $rawHeaders) = $headerParts;
+        [$startLine, $rawHeaders] = $headerParts;
 
         if (preg_match("/(?:^HTTP\/|^[A-Z]+ \S+ HTTP\/)(\d+(?:\.\d+)?)/i", $startLine, $matches) && $matches[1] === '1.0') {
             // Header folding is deprecated for HTTP/1.1, but allowed in HTTP/1.0
@@ -175,10 +171,8 @@ final class Message
      *
      * @param string $path    Path from the start-line
      * @param array  $headers Array of headers (each value an array).
-     *
-     * @return string
      */
-    public static function parseRequestUri($path, array $headers)
+    public static function parseRequestUri(string $path, array $headers): string
     {
         $hostKey = array_filter(array_keys($headers), function ($k) {
             return strtolower($k) === 'host';
@@ -199,10 +193,8 @@ final class Message
      * Parses a request message string into a request object.
      *
      * @param string $message Request message string.
-     *
-     * @return Request
      */
-    public static function parseRequest($message)
+    public static function parseRequest(string $message): RequestInterface
     {
         $data = self::parseMessage($message);
         $matches = [];
@@ -227,10 +219,8 @@ final class Message
      * Parses a response message string into a response object.
      *
      * @param string $message Response message string.
-     *
-     * @return Response
      */
-    public static function parseResponse($message)
+    public static function parseResponse(string $message): ResponseInterface
     {
         $data = self::parseMessage($message);
         // According to https://tools.ietf.org/html/rfc7230#section-3.1.2 the space
@@ -246,7 +236,7 @@ final class Message
             $data['headers'],
             $data['body'],
             explode('/', $parts[0])[1],
-            isset($parts[2]) ? $parts[2] : null
+            $parts[2] ?? null
         );
     }
 }
