@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 use Psr\Http\Message\RequestInterface;
@@ -12,11 +14,9 @@ final class Utils
     /**
      * Remove the items given by the keys, case insensitively from the data.
      *
-     * @param iterable<string> $keys
-     *
-     * @return array
+     * @param string[] $keys
      */
-    public static function caselessRemove($keys, array $data)
+    public static function caselessRemove(array $keys, array $data): array
     {
         $result = [];
 
@@ -25,7 +25,7 @@ final class Utils
         }
 
         foreach ($data as $k => $v) {
-            if (!in_array(strtolower($k), $keys)) {
+            if (!is_string($k) || !in_array(strtolower($k), $keys)) {
                 $result[$k] = $v;
             }
         }
@@ -44,7 +44,7 @@ final class Utils
      *
      * @throws \RuntimeException on error.
      */
-    public static function copyToStream(StreamInterface $source, StreamInterface $dest, $maxLen = -1)
+    public static function copyToStream(StreamInterface $source, StreamInterface $dest, int $maxLen = -1): void
     {
         $bufferSize = 8192;
 
@@ -76,19 +76,16 @@ final class Utils
      * @param int             $maxLen Maximum number of bytes to read. Pass -1
      *                                to read the entire stream.
      *
-     * @return string
-     *
      * @throws \RuntimeException on error.
      */
-    public static function copyToString(StreamInterface $stream, $maxLen = -1)
+    public static function copyToString(StreamInterface $stream, int $maxLen = -1): string
     {
         $buffer = '';
 
         if ($maxLen === -1) {
             while (!$stream->eof()) {
                 $buf = $stream->read(1048576);
-                // Using a loose equality here to match on '' and false.
-                if ($buf == null) {
+                if ($buf === '') {
                     break;
                 }
                 $buffer .= $buf;
@@ -99,8 +96,7 @@ final class Utils
         $len = 0;
         while (!$stream->eof() && $len < $maxLen) {
             $buf = $stream->read($maxLen - $len);
-            // Using a loose equality here to match on '' and false.
-            if ($buf == null) {
+            if ($buf === '') {
                 break;
             }
             $buffer .= $buf;
@@ -120,11 +116,9 @@ final class Utils
      * @param string          $algo      Hash algorithm (e.g. md5, crc32, etc)
      * @param bool            $rawOutput Whether or not to use raw output
      *
-     * @return string Returns the hash of the stream
-     *
      * @throws \RuntimeException on error.
      */
-    public static function hash(StreamInterface $stream, $algo, $rawOutput = false)
+    public static function hash(StreamInterface $stream, string $algo, bool $rawOutput = false): string
     {
         $pos = $stream->tell();
 
@@ -160,10 +154,8 @@ final class Utils
      *
      * @param RequestInterface $request Request to clone and modify.
      * @param array            $changes Changes to apply.
-     *
-     * @return RequestInterface
      */
-    public static function modifyRequest(RequestInterface $request, array $changes)
+    public static function modifyRequest(RequestInterface $request, array $changes): RequestInterface
     {
         if (!$changes) {
             return $request;
@@ -204,13 +196,11 @@ final class Utils
 
         if ($request instanceof ServerRequestInterface) {
             $new = (new ServerRequest(
-                isset($changes['method']) ? $changes['method'] : $request->getMethod(),
+                $changes['method'] ?? $request->getMethod(),
                 $uri,
                 $headers,
-                isset($changes['body']) ? $changes['body'] : $request->getBody(),
-                isset($changes['version'])
-                    ? $changes['version']
-                    : $request->getProtocolVersion(),
+                $changes['body'] ?? $request->getBody(),
+                $changes['version'] ?? $request->getProtocolVersion(),
                 $request->getServerParams()
             ))
             ->withParsedBody($request->getParsedBody())
@@ -226,13 +216,11 @@ final class Utils
         }
 
         return new Request(
-            isset($changes['method']) ? $changes['method'] : $request->getMethod(),
+            $changes['method'] ?? $request->getMethod(),
             $uri,
             $headers,
-            isset($changes['body']) ? $changes['body'] : $request->getBody(),
-            isset($changes['version'])
-                ? $changes['version']
-                : $request->getProtocolVersion()
+            $changes['body'] ?? $request->getBody(),
+            $changes['version'] ?? $request->getProtocolVersion()
         );
     }
 
@@ -241,17 +229,14 @@ final class Utils
      *
      * @param StreamInterface $stream    Stream to read from
      * @param int|null        $maxLength Maximum buffer length
-     *
-     * @return string
      */
-    public static function readLine(StreamInterface $stream, $maxLength = null)
+    public static function readLine(StreamInterface $stream, ?int $maxLength = null): string
     {
         $buffer = '';
         $size = 0;
 
         while (!$stream->eof()) {
-            // Using a loose equality here to match on '' and false.
-            if (null == ($byte = $stream->read(1))) {
+            if ('' === ($byte = $stream->read(1))) {
                 return $buffer;
             }
             $buffer .= $byte;
@@ -294,18 +279,16 @@ final class Utils
      *   buffered and used in subsequent reads.
      *
      * @param resource|string|int|float|bool|StreamInterface|callable|\Iterator|null $resource Entity body data
-     * @param array                                                                  $options  Additional options
-     *
-     * @return StreamInterface
+     * @param array{size?: int, metadata?: array}                                    $options  Additional options
      *
      * @throws \InvalidArgumentException if the $resource arg is not valid.
      */
-    public static function streamFor($resource = '', array $options = [])
+    public static function streamFor($resource = '', array $options = []): StreamInterface
     {
         if (is_scalar($resource)) {
             $stream = self::tryFopen('php://temp', 'r+');
             if ($resource !== '') {
-                fwrite($stream, $resource);
+                fwrite($stream, (string) $resource);
                 fseek($stream, 0);
             }
             return new Stream($stream, $options);
@@ -317,8 +300,9 @@ final class Utils
                  * The 'php://input' is a special stream with quirks and inconsistencies.
                  * We avoid using that stream by reading it into php://temp
                  */
-                $metaData = \stream_get_meta_data($resource);
-                if (isset($metaData['uri']) && $metaData['uri'] === 'php://input') {
+
+                /** @var resource $resource */
+                if ((\stream_get_meta_data($resource)['uri'] ?? '') === 'php://input') {
                     $stream = self::tryFopen('php://temp', 'w+');
                     fwrite($stream, stream_get_contents($resource));
                     fseek($stream, 0);
@@ -326,6 +310,7 @@ final class Utils
                 }
                 return new Stream($resource, $options);
             case 'object':
+                /** @var object $resource */
                 if ($resource instanceof StreamInterface) {
                     return $resource;
                 } elseif ($resource instanceof \Iterator) {
@@ -338,7 +323,7 @@ final class Utils
                         return $result;
                     }, $options);
                 } elseif (method_exists($resource, '__toString')) {
-                    return Utils::streamFor((string) $resource, $options);
+                    return self::streamFor((string) $resource, $options);
                 }
                 break;
             case 'NULL':
@@ -365,21 +350,22 @@ final class Utils
      *
      * @throws \RuntimeException if the file cannot be opened
      */
-    public static function tryFopen($filename, $mode)
+    public static function tryFopen(string $filename, string $mode)
     {
         $ex = null;
-        set_error_handler(function () use ($filename, $mode, &$ex) {
+        set_error_handler(static function (int $errno, string $errstr) use ($filename, $mode, &$ex): bool {
             $ex = new \RuntimeException(sprintf(
                 'Unable to open "%s" using mode "%s": %s',
                 $filename,
                 $mode,
-                func_get_args()[1]
+                $errstr
             ));
 
             return true;
         });
 
         try {
+            /** @var resource $handle */
             $handle = fopen($filename, $mode);
         } catch (\Throwable $e) {
             $ex = new \RuntimeException(sprintf(
@@ -409,11 +395,9 @@ final class Utils
      *
      * @param string|UriInterface $uri
      *
-     * @return UriInterface
-     *
      * @throws \InvalidArgumentException
      */
-    public static function uriFor($uri)
+    public static function uriFor($uri): UriInterface
     {
         if ($uri instanceof UriInterface) {
             return $uri;
