@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 final class Header
@@ -11,10 +13,8 @@ final class Header
      * contains a key, this function will inject a key with a '' string value.
      *
      * @param string|array $header Header to parse into components.
-     *
-     * @return array Returns the parsed header values.
      */
-    public static function parse($header)
+    public static function parse($header): array
     {
         static $trimmed = "\"'  \n\t\r";
         $params = $matches = [];
@@ -44,10 +44,8 @@ final class Header
      * headers into an array of headers with no comma separated values.
      *
      * @param string|array $header Header to normalize.
-     *
-     * @return array Returns the normalized header field values.
      */
-    public static function normalize($header)
+    public static function normalize($header): array
     {
         if (!is_array($header)) {
             return array_map('trim', explode(',', $header));
