@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 use Psr\Http\Message\StreamInterface;
@@ -7,14 +9,12 @@ use Psr\Http\Message\StreamInterface;
 /**
  * Lazily reads or writes to a file that is opened only after an IO operation
  * take place on the stream.
- *
- * @final
  */
-class LazyOpenStream implements StreamInterface
+final class LazyOpenStream implements StreamInterface
 {
     use StreamDecoratorTrait;
 
-    /** @var string File to open */
+    /** @var string */
     private $filename;
 
     /** @var string */
@@ -24,7 +24,7 @@ class LazyOpenStream implements StreamInterface
      * @param string $filename File to lazily open
      * @param string $mode     fopen mode to use when opening the stream
      */
-    public function __construct($filename, $mode)
+    public function __construct(string $filename, string $mode)
     {
         $this->filename = $filename;
         $this->mode = $mode;
@@ -32,10 +32,8 @@ class LazyOpenStream implements StreamInterface
 
     /**
      * Creates the underlying stream lazily when required.
-     *
-     * @return StreamInterface
      */
-    protected function createStream()
+    protected function createStream(): StreamInterface
     {
         return Utils::streamFor(Utils::tryFopen($this->filename, $this->mode));
     }
