@@ -1,15 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 use Psr\Http\Message\StreamInterface;
 
 /**
  * Decorator used to return only a subset of a stream.
- *
- * @final
  */
-class LimitStream implements StreamInterface
+final class LimitStream implements StreamInterface
 {
     use StreamDecoratorTrait;
 
@@ -28,15 +28,15 @@ class LimitStream implements StreamInterface
      */
     public function __construct(
         StreamInterface $stream,
-        $limit = -1,
-        $offset = 0
+        int $limit = -1,
+        int $offset = 0
     ) {
         $this->stream = $stream;
         $this->setLimit($limit);
         $this->setOffset($offset);
     }
 
-    public function eof()
+    public function eof(): bool
     {
         // Always return true if the underlying stream is EOF
         if ($this->stream->eof()) {
@@ -44,7 +44,7 @@ class LimitStream implements StreamInterface
         }
 
         // No limit and the underlying stream is not at EOF
-        if ($this->limit == -1) {
+        if ($this->limit === -1) {
             return false;
         }
 
@@ -53,13 +53,12 @@ class LimitStream implements StreamInterface
 
     /**
      * Returns the size of the limited subset of data
-     * {@inheritdoc}
      */
-    public function getSize()
+    public function getSize(): ?int
     {
         if (null === ($length = $this->stream->getSize())) {
             return null;
-        } elseif ($this->limit == -1) {
+        } elseif ($this->limit === -1) {
             return $length - $this->offset;
         } else {
             return min($this->limit, $length - $this->offset);
@@ -68,9 +67,8 @@ class LimitStream implements StreamInterface
 
     /**
      * Allow for a bounded seek on the read limited stream
-     * {@inheritdoc}
      */
-    public function seek($offset, $whence = SEEK_SET)
+    public function seek($offset, $whence = SEEK_SET): void
     {
         if ($whence !== SEEK_SET || $offset < 0) {
             throw new \RuntimeException(sprintf(
@@ -93,9 +91,8 @@ class LimitStream implements StreamInterface
 
     /**
      * Give a relative tell()
-     * {@inheritdoc}
      */
-    public function tell()
+    public function tell(): int
     {
         return $this->stream->tell() - $this->offset;
     }
@@ -107,7 +104,7 @@ class LimitStream implements StreamInterface
      *
      * @throws \RuntimeException if the stream cannot be seeked.
      */
-    public function setOffset($offset)
+    public function setOffset(int $offset): void
     {
         $current = $this->stream->tell();
 
@@ -132,14 +129,14 @@ class LimitStream implements StreamInterface
      * @param int $limit Number of bytes to allow to be read from the stream.
      *                   Use -1 for no limit.
      */
-    public function setLimit($limit)
+    public function setLimit(int $limit): void
     {
         $this->limit = $limit;
     }
 
-    public function read($length)
+    public function read($length): string
     {
-        if ($this->limit == -1) {
+        if ($this->limit === -1) {
             return $this->stream->read($length);
         }
 
