@@ -1,12 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
+/**
+ * @internal
+ */
 final class Rfc7230
 {
     /**
-     * Header related regular expressions (copied from amphp/http package)
-     * (Note: once we require PHP 7.x we could just depend on the upstream package)
+     * Header related regular expressions (based on amphp/http package)
      *
      * Note: header delimiter (\r\n) is modified to \r?\n to accept line feed only delimiters for BC reasons.
      *
@@ -14,6 +18,6 @@ final class Rfc7230
      *
      * @license https://github.com/amphp/http/blob/v1.0.1/LICENSE
      */
-    const HEADER_REGEX = "(^([^()<>@,;:\\\"/[\]?={}\x01-\x20\x7F]++):[ \t]*+((?:[ \t]*+[\x21-\x7E\x80-\xFF]++)*+)[ \t]*+\r?\n)m";
-    const HEADER_FOLD_REGEX = "(\r?\n[ \t]++)";
+    public const HEADER_REGEX = "(^([^()<>@,;:\\\"/[\]?={}\x01-\x20\x7F]++):[ \t]*+((?:[ \t]*+[\x21-\x7E\x80-\xFF]++)*+)[ \t]*+\r?\n)m";
+    public const HEADER_FOLD_REGEX = "(\r?\n[ \t]++)";
 }
