@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Psr7;
 
 use Psr\Http\Message\StreamInterface;
@@ -7,9 +9,9 @@ use Psr\Http\Message\StreamInterface;
 /**
  * Converts Guzzle streams into PHP stream resources.
  *
- * @final
+ * @see https://www.php.net/streamwrapper
  */
-class StreamWrapper
+final class StreamWrapper
 {
     /** @var resource */
     public $context;
@@ -42,13 +44,11 @@ class StreamWrapper
                 . 'writable, or both.');
         }
 
-        return fopen('guzzle://stream', $mode, null, self::createStreamContext($stream));
+        return fopen('guzzle://stream', $mode, false, self::createStreamContext($stream));
     }
 
     /**
      * Creates a stream context that can be used to open a stream as a php stream resource.
-     *
-     * @param StreamInterface $stream
      *
      * @return resource
      */
@@ -62,14 +62,14 @@ class StreamWrapper
     /**
      * Registers the stream wrapper if needed
      */
-    public static function register()
+    public static function register(): void
     {
         if (!in_array('guzzle', stream_get_wrappers())) {
             stream_wrapper_register('guzzle', __CLASS__);
         }
     }
 
-    public function stream_open($path, $mode, $options, &$opened_path)
+    public function stream_open(string $path, string $mode, int $options, string &$opened_path = null): bool
     {
         $options = stream_context_get_options($this->context);
 
@@ -83,41 +83,48 @@ class StreamWrapper
         return true;
     }
 
-    public function stream_read($count)
+    public function stream_read(int $count): string
     {
         return $this->stream->read($count);
     }
 
-    public function stream_write($data)
+    public function stream_write(string $data): int
     {
-        return (int) $this->stream->write($data);
+        return $this->stream->write($data);
     }
 
-    public function stream_tell()
+    public function stream_tell(): int
     {
         return $this->stream->tell();
     }
 
-    public function stream_eof()
+    public function stream_eof(): bool
     {
         return $this->stream->eof();
     }
 
-    public function stream_seek($offset, $whence)
+    public function stream_seek(int $offset, int $whence): bool
     {
         $this->stream->seek($offset, $whence);
 
         return true;
     }
 
-    public function stream_cast($cast_as)
+    /**
+     * @return resource|false
+     */
+    public function stream_cast(int $cast_as)
     {
         $stream = clone($this->stream);
+        $resource = $stream->detach();
 
-        return $stream->detach();
+        return $resource ?? false;
     }
 
-    public function stream_stat()
+    /**
+     * @return array<int|string, int>
+     */
+    public function stream_stat(): array
     {
         static $modeMap = [
             'r'  => 33060,
@@ -144,7 +151,10 @@ class StreamWrapper
         ];
     }
 
-    public function url_stat($path, $flags)
+    /**
+     * @return array<int|string, int>
+     */
+    public function url_stat(string $path, int $flags): array
     {
         return [
             'dev'     => 0,
